requirements.txt
biopython
argparse
numpy